pyproject.toml
[project]
name = "Atividade-gen-ai-rocket-lab-2026"
version = "1.0.0"
description = "Meu agente text-to-sql"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai==1.84.1",
    "pydantic==2.13.2",
    "python-dotenv==1.2.2",
    "annotated-types==0.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

